with_chainlit_agent/pyproject.toml
[project]
name = "openai-agent-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.4.400",
    "openai-agents>=0.0.9",
    "python-dotenv>=1.1.0",
]
